pyproject.toml
[tool.poetry]
name = "channel-attention"
version = "0.1.0"
description = ""
authors = ["martinwimpff <martin.wimpff@iss.uni-stuttgart.de>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
mne = "^1.5.0"
numpy = "^1.25.2"
scikit-learn = "^1.3.0"
moabb = "^0.5.0"
torch = "^2.0.1"
Braindecode = "^0.7"
PyYAML = "^6.0.1"
pytorch-lightning = "^2.0.8"
einops = "^0.6.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.287"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
